toEqual Jest matcher.
